Minor clean up to remove all warnings that are spurious, so that this builds cleanly on macos now, minus one dependency meson acquires and builds for us (libunibreak)

+++ src/tests/hash/functional.c
@@ -47,7 +47,7 @@ start_one_functest_thread(void *info)
 
     mmm_clean_up_before_exit();
 
-    return (void *)ret;
+    return (void *)(int64_t)ret;
 }
